@@ -1,30 +1,6 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import Home from '@/views/Home.vue'
-import Forms from '@/views/Forms.vue'
-import NotFound from '@/views/NotFound.vue'
+import { createRouter, createWebHistory } from 'vue-router'
+import Home from '../views/Home.vue'
 
-Vue.use(VueRouter)
-
-// window.fbAsyncInit = function () {
-//   FB.init({
-//     appId: '{your-app-id}',
-//     cookie: true,
-//     xfbml: true,
-//     version: '{api-version}'
-//   });
-
-//   FB.AppEvents.logPageView();
-
-// };
-
-// (function (d, s, id) {
-//   var js, fjs = d.getElementsByTagName(s)[0];
-//   if (d.getElementById(id)) { return; }
-//   js = d.createElement(s); js.id = id;
-//   js.src = "https://connect.facebook.net/en_US/sdk.js";
-//   fjs.parentNode.insertBefore(js, fjs);
-// }(document, 'script', 'facebook-jssdk'));
 const routes = [
   {
     path: '/',
@@ -38,14 +14,11 @@ const routes = [
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
     component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-  },
-  { path: '/sign', component: Forms },
-  { path: '*', component: NotFound },
+  }
 ]
 
-const router = new VueRouter({
-  mode: 'history',
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
   routes
 })
 
